Implement image preview and progressbar for txt2img for automatic1111
